+++ node/cli/benches/transaction_pool.rs
@@ -18,10 +18,6 @@
 
 use criterion::{criterion_group, criterion_main, BatchSize, Criterion, Throughput};
 use futures::{future, StreamExt};
-use selendra_node_cli::service::{create_extrinsic, fetch_nonce, FullClient, TransactionPool};
-use selendra_primitives::AccountId;
-use selendra_runtime::{BalancesCall, SudoCall};
-use selendra_runtime_constants::currency::*;
 use sc_client_api::execution_extensions::ExecutionStrategies;
 use sc_service::{
 	config::{
@@ -32,6 +28,10 @@ use sc_service::{
 };
 use sc_transaction_pool::PoolLimit;
 use sc_transaction_pool_api::{TransactionPool as _, TransactionSource, TransactionStatus};
+use selendra_node_cli::service::{create_extrinsic, fetch_nonce, FullClient, TransactionPool};
+use selendra_primitives::AccountId;
+use selendra_runtime::{BalancesCall, SudoCall};
+use selendra_runtime_constants::currency::*;
 use sp_core::{crypto::Pair, sr25519};
 use sp_keyring::Sr25519Keyring;
 use sp_runtime::{generic::BlockId, OpaqueExtrinsic};
