+++ node/executor/benches/bench.rs
@@ -24,8 +24,7 @@ use selendra_node_executor::ExecutorDispatch;
 use selendra_node_test::keyring::*;
 use selendra_primitives::{BlockNumber, Hash};
 use selendra_runtime::{
-	Block, BuildStorage, Call, CheckedExtrinsic, GenesisConfig, Header,
-	UncheckedExtrinsic,
+	Block, BuildStorage, Call, CheckedExtrinsic, GenesisConfig, Header, UncheckedExtrinsic,
 };
 use selendra_runtime_constants::currency::*;
 use sp_core::{
