+++ node/executor/tests/common.rs
@@ -38,8 +38,7 @@ use selendra_node_executor::ExecutorDispatch;
 use selendra_node_test::keyring::*;
 use selendra_primitives::{BlockNumber, Hash};
 use selendra_runtime::{
-	Block, BuildStorage, CheckedExtrinsic, Header, Runtime,
-	UncheckedExtrinsic,
+	Block, BuildStorage, CheckedExtrinsic, Header, Runtime, UncheckedExtrinsic,
 };
 use selendra_runtime_constants::currency::*;
 use sp_externalities::Externalities;
