+++ node/test/src/keyring.rs
@@ -75,7 +75,9 @@ pub fn signed_extra(nonce: Nonce, extra_fee: Balance) -> SignedExtra {
 		frame_system::CheckEra::from(Era::mortal(256, 0)),
 		frame_system::CheckNonce::from(nonce),
 		frame_system::CheckWeight::new(),
-		pallet_transaction_payment::ChargeTransactionPayment::<selendra_runtime::Runtime>::from(extra_fee),
+		pallet_transaction_payment::ChargeTransactionPayment::<selendra_runtime::Runtime>::from(
+			extra_fee,
+		),
 	)
 }
