+++ runtime/common/Cargo.toml
@@ -20,6 +20,7 @@ frame-support = { git = "https://github.com/paritytech/substrate", branch = "pol
 pallet-authorship = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.23", default-features = false }
 pallet-balances = { git = "https://github.com/paritytech/substrate" , branch = "polkadot-v0.9.23", default-features = false }
 pallet-election-provider-multi-phase = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.23", default-features = false }
+pallet-staking = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.23", default-features = false }
 pallet-transaction-payment = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.23", default-features = false }
 pallet-treasury = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.23", default-features = false }
 
@@ -44,6 +45,7 @@ std = [
 	"pallet-authorship/std",
 	"pallet-balances/std",
 	"pallet-election-provider-multi-phase/std",
+	"pallet-staking/std",
 	"pallet-treasury/std",
 	"primitives/std",
 ]
@@ -55,5 +57,6 @@ runtime-benchmarks = [
 try-runtime = [
 	"pallet-authorship/try-runtime",
 	"pallet-authorship/try-runtime",
+	"pallet-staking/try-runtime",
 	"pallet-treasury/try-runtime",
 ]
