+++ runtime/selendra/constants/src/weights/rocksdb_weights.rs
@@ -1,4 +1,3 @@
-
 // This file is part of Substrate.
 
 // Copyright (C) 2022 Parity Technologies (UK) Ltd.
