+++ runtime/selendra/src/config/mod.rs
@@ -1,3 +1,8 @@
+pub mod consensus;
 pub mod core;
-pub mod utility;
+pub mod governance;
+pub mod proxy;
+pub mod staking_support;
 pub mod token_and_relate;
+pub mod treasury;
+pub mod utility;
